Refactor browser layout presenter into modular controllers
Manual testing: not run (container environment).
