+++ src/ui/views/browser/layout/tabSessionController.js
@@ -1,0 +1,14 @@
+export function createTabSessionController({ initTabControls, workspaceManager }) {
+  function init() {
+    initTabControls({
+      onSelectTab: workspaceManager.handleTabSelect,
+      onCloseTab: workspaceManager.handleTabClose
+    });
+  }
+
+  return {
+    init
+  };
+}
+
+export default createTabSessionController;
